order_service/order_service/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
idna==3.10
importlib-metadata==6.11.0
opentelemetry-api==1.32.1
opentelemetry-distro==0.52b1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-django==0.53b1
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-wsgi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
packaging==24.2
pillow==11.1.0
protobuf==5.29.4
requests==2.32.3
setuptools==80.3.1
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
